issue #100: implementation for the restriction of interfaces vcontrold tries to bind to (#101)
* issue #100: testing a dirty implementation for the restriction of the interfaces vcontrold tries to bind to

It features a new "--listen/-L" option for vcontrold to restrict the interface to bind to.
It has to be specified as a local network interface address e.g. 127.0.0.1 or name e.g. localhost.
This option can also be set via xml config.
Default is to listen on all interfaces (as it was).