build workflows via Github Actions (#107)
 Introduces some Github Actions Workflows to build packages for

  *  Linux amd64
  *  Linux armhf (Raspberry Pi)
  *  Linux arm64 (Raspberry Pi)
  *  Windows/Cygwin x86_64
  *  Windows/Cygwin x86

The workflows trigger on push events, as well as when dispatched manually.
When new version tags are introduced, the results of the builds get packaged and offered for download as a new Release on GitHub.

CodeQL workflow now triggers on pushes and pull-requests. It can be used to check for simpler errors and if the code does compile.

Travis CI workflow got removed, as there are no more free CI services.
Appveyor workflow was removed, because of Cygwin binaries are built now via Github Actions.